requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4            # via -r requirements/base.txt, fs, virtualenv
atomicwrites==1.4.0       # via pytest
attrs==19.3.0             # via jsonschema, pytest
aws-sam-translator==1.24.0  # via cfn-lint
aws-xray-sdk==2.6.0       # via moto
bleach==3.1.5             # via -r requirements/base.txt
boto3==1.14.1             # via aws-sam-translator, moto
boto==2.39.0              # via -c requirements/constraints.txt, -r requirements/base.txt, django-pyfs, moto
botocore==1.17.1          # via aws-xray-sdk, boto3, moto, s3transfer
certifi==2020.4.5.2       # via -r requirements/base.txt, requests
cffi==1.14.0              # via cryptography
cfn-lint==0.33.0          # via moto
chardet==3.0.4            # via -r requirements/base.txt, requests
coverage==5.1             # via -r requirements/test.in, pytest-cov
cryptography==2.9.2       # via moto, sshpubkeys
ddt==1.0.0                # via -c requirements/constraints.txt, -r requirements/test.in
decorator==4.4.2          # via networkx
defusedxml==0.6.0         # via -r requirements/base.txt
distlib==0.3.0            # via virtualenv
django-model-utils==4.0.0  # via -r requirements/base.txt, edx-submissions
git+https://github.com/edx/django-pyfs.git@1.0.6#egg=django-pyfs  # via -r requirements/test.in
django-simple-history==2.10.0  # via -r requirements/base.txt
docker==4.2.1             # via moto
docutils==0.15.2          # via botocore
ecdsa==0.15               # via python-jose, sshpubkeys
edx-i18n-tools==0.5.3     # via -r requirements/base.txt
edx-submissions==3.1.10    # via -r requirements/base.txt
factory-boy==2.12.0       # via -r requirements/test.in
faker==4.1.0              # via factory-boy
filelock==3.0.12          # via tox, virtualenv
freezegun==0.3.14         # via -r requirements/test.in
fs==2.0.18                # via -c requirements/constraints.txt, -r requirements/base.txt, django-pyfs, xblock
future==0.18.2            # via aws-xray-sdk
html5lib==1.0.1           # via -r requirements/base.txt
idna==2.8                 # via -c requirements/constraints.txt, -r requirements/base.txt, moto, requests
importlib-metadata==1.6.1  # via -r requirements/base.txt, importlib-resources, jsonpickle, jsonschema, path, pluggy, tox, virtualenv
importlib-resources==1.5.0  # via cfn-lint, virtualenv
jinja2==2.11.2            # via moto
jmespath==0.10.0          # via boto3, botocore
jsondiff==1.1.2           # via moto
jsonfield2==3.0.3         # via -c requirements/constraints.txt, -r requirements/base.txt, edx-submissions
jsonpatch==1.25           # via cfn-lint
jsonpickle==1.4.1         # via aws-xray-sdk
jsonpointer==2.0          # via jsonpatch
jsonschema==3.2.0         # via aws-sam-translator, cfn-lint
junit-xml==1.9            # via cfn-lint
lazy==1.4                 # via -r requirements/base.txt
libsass==0.20.0           # via -r requirements/base.txt
loremipsum==1.0.5         # via -c requirements/constraints.txt, -r requirements/base.txt
lxml==4.5.1               # via -r requirements/base.txt, xblock
markupsafe==1.1.1         # via -r requirements/base.txt, jinja2, xblock
mock==3.0.5               # via -c requirements/constraints.txt, -r requirements/test.in, moto
more-itertools==8.3.0     # via -r requirements/base.txt, -r requirements/test.in, pytest, zipp
moto==1.3.14              # via -r requirements/test.in
networkx==2.4             # via cfn-lint
packaging==20.4           # via -r requirements/base.txt, bleach, tox
path.py==12.4.0           # via -r requirements/base.txt, edx-i18n-tools
path==13.1.0              # via -r requirements/base.txt, path.py
pathlib2==2.3.5           # via pytest
pluggy==0.13.1            # via pytest, tox
polib==1.1.0              # via -r requirements/base.txt, edx-i18n-tools
py==1.8.1                 # via pytest, tox
pyasn1==0.4.8             # via python-jose, rsa
pycparser==2.20           # via cffi
pyparsing==2.4.7          # via -r requirements/base.txt, packaging
pyrsistent==0.16.0        # via jsonschema
pytest-cov==2.7.1         # via -r requirements/test.in
pytest-django==3.7.0      # via -r requirements/test.in
pytest==4.5.0             # via -r requirements/test.in, pytest-cov, pytest-django
python-dateutil==2.4.0    # via -c requirements/constraints.txt, -r requirements/base.txt, botocore, faker, freezegun, moto, xblock
python-jose==3.1.0        # via moto
python-swiftclient==3.9.0  # via -c requirements/constraints.txt, -r requirements/base.txt
pytz==2020.1              # via -r requirements/base.txt, django, edx-submissions, fs, moto, xblock
pyyaml==5.3.1             # via -r requirements/base.txt, cfn-lint, edx-i18n-tools, moto, xblock
requests==2.23.0          # via -r requirements/base.txt, docker, moto, python-swiftclient, responses
responses==0.10.15        # via moto
rsa==4.2                  # via python-jose
s3transfer==0.3.3         # via boto3
six==1.15.0               # via -r requirements/base.txt, aws-sam-translator, bleach, cfn-lint, cryptography, django-pyfs, django-simple-history, docker, ecdsa, edx-i18n-tools, freezegun, fs, html5lib, jsonschema, junit-xml, libsass, mock, moto, packaging, pathlib2, pytest, python-dateutil, python-jose, python-swiftclient, responses, tox, virtualenv, websocket-client, xblock
sqlparse==0.3.1           # via -r requirements/base.txt, django
sshpubkeys==3.1.0         # via moto
testfixtures==6.14.1      # via -r requirements/test.in
text-unidecode==1.3       # via faker
toml==0.10.1              # via tox
tox==3.15.2               # via -r requirements/test.in
urllib3==1.25.9           # via -r requirements/base.txt, botocore, requests
virtualenv==20.0.21       # via tox
voluptuous==0.11.7        # via -c requirements/constraints.txt, -r requirements/base.txt
wcwidth==0.2.4            # via pytest
web-fragments==0.3.2      # via -r requirements/base.txt, xblock
webencodings==0.5.1       # via -r requirements/base.txt, bleach, html5lib
webob==1.8.6              # via -r requirements/base.txt, xblock
websocket-client==0.57.0  # via docker
werkzeug==1.0.1           # via moto
wrapt==1.12.1             # via aws-xray-sdk
xblock-sdk==0.2.0         # via -r requirements/base.txt
xblock==1.3.1             # via -r requirements/base.txt
xmltodict==0.12.0         # via moto
zipp==1.0.0               # via -c requirements/constraints.txt, -r requirements/base.txt, importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
